requirements.txt
Django>=5.2,<5.3
wagtail>=7.2,<7.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
